# feat(integer): add scalar comparisons