--- a/tfhe/src/integer/server_key/radix_parallel/mod.rs
+++ b/tfhe/src/integer/server_key/radix_parallel/mod.rs
@@ -5,6 +5,7 @@ mod mul;
 mod neg;
 mod rotate;
 mod scalar_add;
+mod scalar_comparison;
 mod scalar_mul;
 mod scalar_rotate;
 mod scalar_shift;