This adds scalar comparisons and min/max

- eq (==) and ne (!=) have similar performances to the non scalar version, only gaining a few milliseconds when the scalar value is smaller than the encrypted value

- orderings (<, >, <= ,>=) have more interesting performances gain the more the scalar value is small than the number of encrypted bits in the integer ciphertext.

  e.g: comparing an encrypted U256 to a value <= u128::MAX brings the comparison time from 234 ms to 194 ms

  comparing an encrypted U256 to a value <= u64::MAX brings the comparison time from 234 ms to 169 ms